+++ spec/truffle/interop/equal_spec.rb
@@ -1,4 +1,4 @@
-# Copyright (c) 2017 Oracle and/or its affiliates. All rights reserved. This
+# Copyright (c) 2017, 2018 Oracle and/or its affiliates. All rights reserved. This
 # code is released under a tri EPL/GPL/LGPL license. You can use it,
 # redistribute it and/or modify it under the terms of the:
 #
@@ -9,16 +9,24 @@
 require_relative '../../ruby/spec_helper'
 
 describe "Calling #equal? on a foreign object" do
+  
   it "tests reference equality" do
     a = Truffle::Debug.foreign_object
-    a.equal?(a).should == true
+    a.equal?(a).should be_true
     b = Truffle::Debug.foreign_object
-    a.equal?(b).should == false
+    a.equal?(b).should be_false
   end
 
-  it "returns false for other objects" do
-    foreign = Truffle::Debug.foreign_object
-    foreign.equal?(Object.new).should == false
-    foreign.equal?(42).should == false
+  guard -> { !Truffle.native? } do
+    it "looks at the underlying object for Java interop" do
+      big_integer = Truffle::Interop.java_type("java.math.BigInteger")
+      a = big_integer.ONE
+      b = big_integer.ONE
+      c = big_integer.TEN
+      a.equal?(a).should be_true
+      a.equal?(b).should be_true
+      a.equal?(c).should be_false
+    end
   end
+  
 end